+++ docs/.vitepress/theme/NavTitle.vue
@@ -1,6 +1,6 @@
 <template>
   <span class="sjmcl-nav-title">
-    <span>SJMC</span>
+    <span>Amethyst</span>
     <span class="highlight">&nbsp;L</span>
     <span>auncher</span>
   </span>