+++ docs/zh-Hans/docs/changelog.md
@@ -10,5 +10,5 @@
 - 移除 SJMC 用户中心登录组件
 - 移除发现页社区新闻功能
 - 仅保留简体中文和繁体中文
-- 软件更名为 ArcMC Launcher
+- 软件更名为 Amethyst Launcher
 - 目前仅支持 Windows 平台